Update API URLs and refactor register remote data source
* Change API base URL from 'http://10.0.2.2/grd' to 'http://192.168.1.17
