+++ lib/core/Services/Api/app_url.dart
@@ -3,7 +3,7 @@ import 'dart:convert';
 import 'package:http/http.dart' as http;
 
 class ApiUrl {
-  static const String baseUrl = 'http://10.0.2.2/grd';
+  static const String baseUrl = 'http://192.168.1.17/grd';
 
   static const String loginUrl = '$baseUrl/auth/login.php';
   static const String registerUrl = '$baseUrl/auth/signup.php';
@@ -15,19 +15,4 @@ class ApiUrl {
       '$baseUrl/medical_information.php';
   static const String forgetPasswordUrl = '$baseUrl/forgot_password.php';
   static const String resetPasswordUrl = '$baseUrl/reset_password.php';
-
-  static Future<http.MultipartRequest> postData(
-      {required String endPoint}) async {
-    var response = http.MultipartRequest("Post", Uri.parse(endPoint));
-    return response;
-  }
-
-  static Future<http.Response> postRequest(
-      {required String body,
-      required Map<String, String> header,
-      required String endPoint}) async {
-    var response =
-        await http.post(Uri.parse(endPoint), body: body, headers: header);
-    return response;
-  }
 }
